making a resolve user util function and using it in add leader intent handlers
