+++ lambda/custom/handlers/Channels/AddLeaderIntentHandlers.js
@@ -1,6 +1,37 @@
 const Alexa = require('ask-sdk-core');
 const { ri } = require('@jargon/alexa-skill-sdk');
-const { login, resolveUsername, resolveChannelname, addLeader } = require('../../helperFunctions');
+const { login, addLeader } = require('../../helperFunctions');
+const { resolveChannel, resolveUser } = require('../../utils');
+
+const StartedAddLeaderIntentHandler = {
+	canHandle(handlerInput) {
+		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
+        handlerInput.requestEnvelope.request.intent.name === 'AddLeaderIntent' &&
+        handlerInput.requestEnvelope.request.dialogState === 'STARTED';
+	},
+	handle(handlerInput) {
+		const currentIntent = handlerInput.requestEnvelope.request.intent;
+		const intentSlots = currentIntent.slots;
+
+		const { attributesManager } = handlerInput;
+		const sessionAttributes = attributesManager.getSessionAttributes() || {};
+
+		delete sessionAttributes.user;
+		delete sessionAttributes.channel;
+		delete sessionAttributes.similarusers;
+		delete sessionAttributes.similarChannels;
+
+		if (intentSlots.username.confirmationStatus === 'NONE' && intentSlots.username.value) {
+			return resolveUser(handlerInput);
+		} else if (intentSlots.channelname.confirmationStatus === 'NONE' && intentSlots.channelname.value) {
+			return resolveChannel(handlerInput);
+		}
+
+		return handlerInput.responseBuilder
+			.addDelegateDirective(currentIntent)
+			.getResponse();
+	},
+};
 
 const AddLeaderIntentHandler = {
 	canHandle(handlerInput) {
@@ -48,188 +79,6 @@ const DeniedAddLeaderIntentHandler = {
 	},
 };
 
-const UnconfirmedLeaderAddLeaderIntentHandler = {
-	canHandle(handlerInput) {
-		return Alexa.getRequestType(handlerInput.requestEnvelope) === 'IntentRequest'
-            && handlerInput.requestEnvelope.request.intent.name === 'AddLeaderIntent'
-            && handlerInput.requestEnvelope.request.intent.slots.username.confirmationStatus === 'NONE'
-            && handlerInput.requestEnvelope.request.intent.slots.username.value;
-	},
-	async handle(handlerInput) {
-		const updatedIntent = handlerInput.requestEnvelope.request.intent;
-		const updatedSlots = updatedIntent.slots;
-
-		const { attributesManager } = handlerInput;
-		const sessionAttributes = attributesManager.getSessionAttributes() || {};
-
-		if (updatedSlots.choice && updatedSlots.choice.value) {
-			// get the array of users which the user was asked for
-			const users = sessionAttributes.similarusers;
-
-			// if the user selects an invalid choice then ask for an appropriate choice
-			if (Number(updatedSlots.choice.value) === 0 || Number(updatedSlots.choice.value) > users.length) {
-				let users_list = '';
-				for (const { name } of users) {
-					users_list += `${ name }, `;
-				}
-				const speechText = ri('RESOLVE_USERNAME.ASK_CHOICE', { choice_limit: users.length, users_list });
-				const slotName = 'choice';
-				return handlerInput.jrb
-					.speak(speechText)
-					.reprompt(speechText)
-					.addElicitSlotDirective(slotName)
-					.getResponse();
-			}
-
-			// if everything is correct then proceed to ask for confirmation
-			const userDetails = users[Number(updatedSlots.choice.value) - 1];
-			updatedSlots.username.value = userDetails.name;
-			sessionAttributes.user = userDetails;
-			delete updatedSlots.choice.value;
-			return handlerInput.responseBuilder
-				.addDelegateDirective(updatedIntent)
-				.getResponse();
-		}
-
-		if (updatedSlots.username.value) {
-			const {
-				accessToken,
-			} = handlerInput.requestEnvelope.context.System.user;
-
-			const headers = await login(accessToken);
-
-			// get the array of similar usernames
-			const users = await resolveUsername(updatedSlots.username.value, headers);
-
-			if (users.length === 0) {
-				const speechText = ri('RESOLVE_USERNAME.NO_USER', { username: updatedSlots.username.value });
-				const repromptText = ri('GENERIC_REPROMPT');
-				return handlerInput.jrb
-					.speak(speechText)
-					.speak(repromptText)
-					.reprompt(repromptText)
-					.getResponse();
-				// if there's only one similar channel, then change the slot value to the matching channel
-			} else if (users.length === 1) {
-				updatedSlots.username.value = users[0].name;
-				sessionAttributes.user = users[0];
-			} else {
-				sessionAttributes.similarusers = users;
-				let user_names = '';
-				for (const { name } of users) {
-					user_names += `${ name }, `;
-				}
-
-				const speechText = ri('RESOLVE_USERNAME.SIMILAR_USERS', { user_names });
-				const slotName = 'choice';
-
-
-				return handlerInput.jrb
-					.speak(speechText)
-					.reprompt(speechText)
-					.addElicitSlotDirective(slotName)
-					.getResponse();
-			}
-		}
-
-		return handlerInput.responseBuilder
-			.addDelegateDirective(updatedIntent)
-			.getResponse();
-
-	},
-};
-
-const UnconfirmedChannelAddLeaderIntentHandler = {
-	canHandle(handlerInput) {
-		return Alexa.getRequestType(handlerInput.requestEnvelope) === 'IntentRequest'
-            && handlerInput.requestEnvelope.request.intent.name === 'AddLeaderIntent'
-            && handlerInput.requestEnvelope.request.intent.slots.channelname.confirmationStatus === 'NONE'
-            && handlerInput.requestEnvelope.request.intent.slots.channelname.value;
-	},
-	async handle(handlerInput) {
-		const updatedIntent = handlerInput.requestEnvelope.request.intent;
-		const updatedSlots = updatedIntent.slots;
-
-		const { attributesManager } = handlerInput;
-		const sessionAttributes = attributesManager.getSessionAttributes() || {};
-
-		if (updatedSlots.choice && updatedSlots.choice.value) {
-			// get the array of users which the user was asked for
-			const channels = sessionAttributes.similarchannels;
-
-			// if the user selects an invalid choice then ask for an appropriate choice
-			if (Number(updatedSlots.choice.value) === 0 || Number(updatedSlots.choice.value) > channels.length) {
-				let channels_list = '';
-				for (const { name } of channels) {
-					channels_list += `${ name }, `;
-				}
-				const speechText = ri('RESOLVE_CHANNEL.ASK_CHOICE', { choice_limit: channels.length, channels_list });
-				const slotName = 'choice';
-				return handlerInput.jrb
-					.speak(speechText)
-					.reprompt(speechText)
-					.addElicitSlotDirective(slotName)
-					.getResponse();
-			}
-
-			// if everything is correct then proceed to ask for confirmation
-			const userDetails = channels[Number(updatedSlots.choice.value) - 1];
-			updatedSlots.channelname.value = userDetails.name;
-			sessionAttributes.channel = userDetails;
-			delete updatedSlots.choice.value;
-			return handlerInput.responseBuilder
-				.addDelegateDirective(updatedIntent)
-				.getResponse();
-		}
-
-		if (updatedSlots.channelname.value) {
-			const {
-				accessToken,
-			} = handlerInput.requestEnvelope.context.System.user;
-
-			const headers = await login(accessToken);
-
-			// get the array of similar usernames
-			const channels = await resolveChannelname(updatedSlots.channelname.value, headers);
-
-			if (channels.length === 0) {
-				const speechText = ri('RESOLVE_CHANNEL.NO_CHANNEL', { channel_name: updatedSlots.channelname.value });
-				const repromptText = ri('GENERIC_REPROMPT');
-				return handlerInput.jrb
-					.speak(speechText)
-					.speak(repromptText)
-					.reprompt(repromptText)
-					.getResponse();
-				// if there's only one similar channel, then change the slot value to the matching channel
-			} else if (channels.length === 1) {
-				updatedSlots.channelname.value = channels[0].name;
-				sessionAttributes.channel = channels[0];
-			} else {
-				sessionAttributes.similarchannels = channels;
-				let channel_names = '';
-				for (const { name } of channels) {
-					channel_names += `${ name }, `;
-				}
-
-				const speechText = ri('RESOLVE_CHANNEL.SIMILAR_CHANNELS', { channel_names });
-				const slotName = 'choice';
-
-
-				return handlerInput.jrb
-					.speak(speechText)
-					.reprompt(speechText)
-					.addElicitSlotDirective(slotName)
-					.getResponse();
-			}
-		}
-
-		return handlerInput.responseBuilder
-			.addDelegateDirective(updatedIntent)
-			.getResponse();
-
-	},
-};
-
 const InProgressAddLeaderIntentHandler = {
 	canHandle(handlerInput) {
 		return Alexa.getRequestType(handlerInput.requestEnvelope) === 'IntentRequest'
@@ -237,18 +86,24 @@ const InProgressAddLeaderIntentHandler = {
             && handlerInput.requestEnvelope.request.dialogState !== 'COMPLETED';
 	},
 	handle(handlerInput) {
-		const updatedIntent = handlerInput.requestEnvelope.request.intent;
+		const { intent } = handlerInput.requestEnvelope.request;
+		const intentSlots = intent.slots;
+
+		if (intentSlots.username.confirmationStatus === 'NONE' && intentSlots.username.value) {
+			return resolveUser(handlerInput);
+		} else if (intentSlots.channelname.confirmationStatus === 'NONE' && intentSlots.channelname.value) {
+			return resolveChannel(handlerInput);
+		}
 
 		return handlerInput.responseBuilder
-			.addDelegateDirective(updatedIntent)
+			.addDelegateDirective(intent)
 			.getResponse();
 	},
 };
 
 module.exports = {
+	StartedAddLeaderIntentHandler,
 	AddLeaderIntentHandler,
 	DeniedAddLeaderIntentHandler,
-	UnconfirmedLeaderAddLeaderIntentHandler,
-	UnconfirmedChannelAddLeaderIntentHandler,
 	InProgressAddLeaderIntentHandler,
 };
